+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Tx.hs
@@ -13,6 +13,7 @@ module Cardano.Ledger.Shelley.Tx (
   sizeShelleyTxF,
   wireSizeShelleyTxF,
   segWitAnnTx,
+  segWitTx,
   mkBasicShelleyTx,
   shelleyMinFeeTx,
   witsFromTxWitnesses,